+++ server/libs/modules/components/ai/llm/open-router/src/main/java/com/bytechef/component/ai/llm/open/router/action/OpenRouterChatAction.java
@@ -64,7 +64,7 @@ import static com.bytechef.component.definition.Authorization.TOKEN;
 import static com.bytechef.component.definition.ComponentDsl.action;
 
 import com.bytechef.component.ai.llm.ChatModel;
-import com.bytechef.component.ai.llm.open.router.util.OpenRouterChatModel;
+import com.bytechef.component.ai.llm.open.router.model.OpenRouterChatModel;
 import com.bytechef.component.ai.llm.util.ModelUtils;
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
