+++ server/libs/modules/components/ai/llm/open-router/src/main/java/com/bytechef/component/ai/llm/open/router/model/OpenRouterChatModel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.ai.llm.open.router.util;
+package com.bytechef.component.ai.llm.open.router.model;
 
 import com.bytechef.component.ai.llm.util.ModelUtils;
 import java.util.ArrayList;
